data_collection/requirements.txt
numpy
pandas
huggingface
huggingface_hub
transformers
datasets
tokenizers
python-dotenv
python-soundfile
librosa
tqdm
kagglehub